Use simpler approach to guard against back-pressure
I'm not sure if the previous approach would call the `done` callback
twice if `req.write()` returned false. Also, there was a bug with
Node.js v0.11 where an error was thrown because of the callback given to
the `req.write()` method. I'm not sure if this was a bug in Node.js
v0.11.9 or my code, but this new code doesn't rely on the callback and
hence also works in Node.js v0.11.